FEAT Firestore object to POJO works

+++ app/src/main/java/com/example/listapp/model/GlassDoor.java
@@ -17,7 +17,7 @@ public class GlassDoor extends Door {
     }
 
     public GlassDoor(int id, int weight, int viewCount, float price,
-                     List<Integer> dimensions, List<String> name, String description,
+                     List<Long> dimensions, List<String> name, String description,
                      List<String> colour, List<String> image) {
         this.id = id;
         this.weight = weight;
